apply padding / fix rerender
