+++ src/components/Molecules/Board/slick.css
@@ -53,3 +53,9 @@
   color: #476cff;
   height: 6px;
 }
+
+.slick-slide {
+  width: 100%;
+  padding: 0 8px;
+  box-sizing: border-box;
+}
